Factored out ParseOptions and its method outside parse.
Now parse_ledger function receives options as arg,
and use it internally.

Those logics are now handled in parse::adaptor,
which can be used for single entry parse logic later.
